+++ src/constants/wavesColors.json
